style: move CardStatesType
