--- src/pages/CardCreator/types/index.ts
@@ -1,2 +1,0 @@
-export * from './types';
-export * from './errorTypes';
