+++ src/stores/CardCreatorContext/CardCreatorStates/types.ts
@@ -9,3 +9,5 @@ export type ExpireDatesState = Omit<InputStateType<ExpireDate>, 'type'>[];
 export type SecurityCodesState = Omit<InputStateType<SecurityCode>, 'placeholder'>[];
 
 export type PasswordsState = Omit<InputStateType<Password>, 'type' | 'placeholder'>[];
+
+export type CardNicknameState = Omit<InputStateType<Password>, 'type' | 'placeholder' | 'checkIsAllowInput'>;
